Add a min_element_by implementation and use it
The implementation of `nearest_neighbor` using `min_element` with a
comparator called the `distance` twice per comparison, with no
memoization, which is inefficient because the same point is likely to be
compared many times, but only needs to be calculated once.

This could still be solved using a comparator, if I knew how to memoize
things easily. Perhaps I'll look into that.